rename module hackers into applications
