+++ applications/static/css/profile.css
