+++ applications/static/css/vote.css
