+++ applications/static/degrees.json
