+++ applications/static/schools.json
