+++ applications/templates/mails/confirmation_message.html
