+++ applications/templates/mails/confirmation_message.txt
