+++ applications/templates/mails/invitation_message.html
