+++ applications/templates/mails/invitation_message.txt
