+++ applications/templates/mails/invitation_subject.txt
